+++ app/src/main/java/com/mashup/ipdam/home/HomeFragment.kt
@@ -1,16 +1,51 @@
 package com.mashup.ipdam.home
 
+import android.graphics.PointF
+import android.util.Log
 import com.mashup.base.BaseFragment
 import com.mashup.ipdam.R
 import com.mashup.ipdam.databinding.FragmentHomeBinding
+import com.naver.maps.map.NaverMap
+import com.naver.maps.map.OnMapReadyCallback
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
+class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home), OnMapReadyCallback {
 
-    override var logTag: String = "MapFragment"
+    override var logTag: String = "HomeFragment"
 
-    companion object {
-        fun getInstance() = HomeFragment()
+    private lateinit var map: NaverMap
+
+    override fun initLayout() {
+        binding.map.getMapAsync(this)
+    }
+
+    private fun getMapBoundsOnScreen() {
+        val locationOnScreen = getMapLocation()
+        val topLeftPointF = PointF(0f, 0f)
+        val bottomLeftPointF = PointF(0f, locationOnScreen[1].toFloat())
+        val topRightPointF = PointF(locationOnScreen[0].toFloat(), 0f)
+        val bottomRightPointF = PointF(locationOnScreen[0].toFloat(), locationOnScreen[1].toFloat())
+
+        val projection = map.projection
+        val topLeftCoord = projection.fromScreenLocation(topLeftPointF)
+        val topRightCoord = projection.fromScreenLocation(topRightPointF)
+        val bottomLeftCoord = projection.fromScreenLocation(bottomLeftPointF)
+        val bottomRightCoord = projection.fromScreenLocation(bottomRightPointF)
+
+        Log.d(logTag, "topLeft: $topLeftCoord topRight: $topRightCoord" +
+                "bottomLeft: $bottomLeftCoord bottomRight: $bottomRightCoord")
+    }
+
+    /**
+     * @return IntArray : 절대 좌표 x, y
+     */
+    private fun getMapLocation() = intArrayOf(0, 0).apply {
+        binding.map.getLocationOnScreen(this)
+    }
+
+    override fun onMapReady(map: NaverMap) {
+        this.map = map
+        getMapBoundsOnScreen()
     }
 }
