+++ app/src/main/res/values/dimens.xml
@@ -2,4 +2,6 @@
 <resources>
     <dimen name="radius_corner">8dp</dimen>
     <dimen name="height_search">40dp</dimen>
+    <dimen name="size_navigation_icon">20dp</dimen>
+    <dimen name="size_add_button">50dp</dimen>
 </resources>
